Ignore files with specific keywords
Do not commit files that contain private (to keep passwords and keys safe) or notes, test, todelete (to avoid having my notes, tests and temp scripts all over the place).
I also excluded .pyc files as they may contain the passwords and keys too.
I also excluded the xlsx as they are not supposed to be part of the project as of now. I am only using xlsx now to test the logic before writing it in Python.